added logout button functionality, after logging out, upon re-signing in, navigating to other tabs will cause the react router to not display the dashboard. Need to look in to this
