+++ src/main/index.js
@@ -2,8 +2,8 @@
 import { app, BrowserWindow, ipcMain } from "electron";
 import * as path from "path";
 import { format as formatUrl } from "url";
-import server from "../../server/server.js";
-
+// import server from "../../server/server.js";
+import installExtension, { REDUX_DEVTOOLS, REACT_DEVELOPER_TOOLS } from 'electron-devtools-installer';
 
 import verifyCode from "./twilio/verifyCode";
 import verifyMobileNumber from "./twilio/verifyMobile";
@@ -76,6 +76,11 @@ app.on("ready", () => {
   mainWindow = createMainWindow();
 });
 
+app.whenReady().then(() => {
+  installExtension([REDUX_DEVTOOLS, REACT_DEVELOPER_TOOLS])
+      .then((name) => console.log(`Added Extension:  ${name}`))
+      .catch((err) => console.log('An error occurred: ', err));
+});
 // if (module.hot) {
 //   module.hot.accept();
 // }
