+++ src/renderer/index.html
@@ -7,7 +7,6 @@
     <title>Document</title>
 </head>
 <body>
-    <h1>hello from indexhtml</h1>
     <div id="app"></div>
 </body>
 </html>
